requirements.txt
playwright
beautifulsoup4
Pillow
